Made a just in case check in BeamRoutine more info ->
The BeamRoutine now first checks if the beam is on cooldown, if not then it proceeds to the actual beam actions.

The if else statement also now always checks if the beam is on cooldown instead of just proceeding to the onCooldown actions without actually checking if the beam is on cooldown.